+++ commands/staff/settings.js
@@ -9,48 +9,48 @@ module.exports = {
         .setDescription('Wyświetla lub edytuj ustawienia bota.')
         .setContexts(InteractionContextType.Guild)
         .setDefaultMemberPermissions(PermissionFlagsBits.Administrator)
-        .addSubcommand(sub =>
-            sub.setName('view')
-                .setDescription('Wyświetla aktywne ustawienia.')
+        .addSubcommand(sub => sub
+            .setName('view')
+            .setDescription('Wyświetla aktywne ustawienia.')
         )
-        .addSubcommand(sub =>
-            sub.setName('nickname')
-                .setDescription('Ustawia lub usuwa rolę zmiany pseudonimu.')
-                .addRoleOption(option =>
-                    option.setName('rola')
-                        .setDescription('Rola zmiany pseudonimu.')
-                        .setRequired(false)
-                )
+        .addSubcommand(sub => sub
+            .setName('nickname')
+            .setDescription('Ustawia lub usuwa rolę zmiany pseudonimu.')
+            .addRoleOption(option => option
+                .setName('rola')
+                .setDescription('Rola zmiany pseudonimu.')
+                .setRequired(false)
+            )
         )
-        .addSubcommand(sub =>
-            sub.setName('snitch')
-                .setDescription('Ustawia lub usuwa kanał systemu zgłoszeń.')
-                .addChannelOption(option =>
-                    option.setName('kanał')
-                        .setDescription('Kanał zgłoszeń.')
-                        .setRequired(false)
-                        .addChannelTypes(ChannelType.GuildText, ChannelType.GuildAnnouncement, ChannelType.AnnouncementThread)
-                )
+        .addSubcommand(sub => sub
+            .setName('snitch')
+            .setDescription('Ustawia lub usuwa kanał systemu zgłoszeń.')
+            .addChannelOption(option => option
+                .setName('kanał')
+                .setDescription('Kanał zgłoszeń.')
+                .setRequired(false)
+                .addChannelTypes(ChannelType.GuildText, ChannelType.GuildAnnouncement, ChannelType.AnnouncementThread)
+            )
         )
-        .addSubcommand(sub =>
-            sub.setName('memes')
-                .setDescription('Ustawia lub usuwa kanał z memami (automatyczne reakcje).')
-                .addChannelOption(option =>
-                    option.setName('kanał')
-                        .setDescription('Kanał memy.')
-                        .setRequired(false)
-                        .addChannelTypes(ChannelType.GuildText, ChannelType.GuildAnnouncement, ChannelType.AnnouncementThread)
-                )
+        .addSubcommand(sub => sub
+            .setName('memes')
+            .setDescription('Ustawia lub usuwa kanał z memami (automatyczne reakcje).')
+            .addChannelOption(option => option
+                .setName('kanał')
+                .setDescription('Kanał memy.')
+                .setRequired(false)
+                .addChannelTypes(ChannelType.GuildText, ChannelType.GuildAnnouncement, ChannelType.AnnouncementThread)
+            )
         )
-        .addSubcommand(sub =>
-            sub.setName('statistics')
-                .setDescription('Ustawia lub usuwa kanał do wyświetlania statystyk.')
-                .addChannelOption(option =>
-                    option.setName('kanał')
-                        .setDescription('Kanał statystyk.')
-                        .setRequired(false)
-                        .addChannelTypes(ChannelType.GuildText, ChannelType.GuildAnnouncement, ChannelType.AnnouncementThread)
-                )
+        .addSubcommand(sub => sub
+            .setName('statistics')
+            .setDescription('Ustawia lub usuwa kanał do wyświetlania statystyk.')
+            .addChannelOption(option => option
+                .setName('kanał')
+                .setDescription('Kanał statystyk.')
+                .setRequired(false)
+                .addChannelTypes(ChannelType.GuildText, ChannelType.GuildAnnouncement, ChannelType.AnnouncementThread)
+            )
         ),
     async execute(interaction, logger) {
         const { utils } = interaction.client;
@@ -136,7 +136,7 @@ module.exports = {
                             logger.info(`[Slash ▸ Settings] Snitch channel removed for '${interaction.guild.id}'.`);
                             return await utils.reply.success(interaction, 'CHANNEL_REMOVED');
                         } else {
-                            return await utils.reply.error(interaction, 'CHANNE_NOT_SET');
+                            return await utils.reply.error(interaction, 'CHANNEL_NOT_SET');
                         }
                     }
 
@@ -168,7 +168,7 @@ module.exports = {
                             logger.info(`[Slash ▸ Settings] Memes channel removed for '${interaction.guild.id}'.`);
                             return await utils.reply.success(interaction, 'CHANNEL_REMOVED');
                         } else {
-                            return await utils.reply.error(interaction, 'CHANNE_NOT_SET');
+                            return await utils.reply.error(interaction, 'CHANNEL_NOT_SET');
 
                         }
                     }
@@ -203,7 +203,7 @@ module.exports = {
                             return await utils.reply.success(interaction, 'CHANNEL_REMOVED');
 
                         } else {
-                            return await utils.reply.error(interaction, 'CHANNE_NOT_SET');
+                            return await utils.reply.error(interaction, 'CHANNEL_NOT_SET');
                         }
                     }
 
@@ -230,7 +230,7 @@ module.exports = {
                     await utils.reply.error(interaction, 'PARAMETER_NOT_FOUND');
             }
         } catch (err) {
-            logger.error(`[Slash ▸ Settings] An error occurred for '${interaction.guild.id}':\n${err}`);
+            logger.error(`[Slash ▸ Settings] An error occurred in subcommand '${subcommand}' for '${interaction.guild.id}':\n${err}`);
             await utils.reply.error(interaction, 'SETTINGS_ERROR');
         }
     },
