+++ commands/utility/avatar.js
@@ -8,10 +8,10 @@ module.exports = {
         .setName('avatar')
         .setDescription('Wyświetla avatar użytkownika.')
         .setContexts(InteractionContextType.Guild)
-        .addUserOption(option =>
-            option.setName('użytkownik')
-                .setDescription('Wybierz użytkownika, którego avatar chcesz zobaczyć.')
-                .setRequired(false)
+        .addUserOption(option => option
+            .setName('użytkownik')
+            .setDescription('Wybierz użytkownika, którego avatar chcesz zobaczyć.')
+            .setRequired(false)
         ),
     async execute(interaction) {
         const { utils } = interaction.client;
