+++ lib/services/packageUpdater.js
@@ -24,8 +24,8 @@ function packageUpdater(logger) {
 
             try {
                 outdatedPackages = JSON.parse(output || '{}');
-            } catch (parseErr) {
-                logger.error(`[Client] Error parsing 'pnpm outdated' output:\n${parseErr.message}`);
+            } catch (err) {
+                logger.error(`[Client] Error parsing 'pnpm outdated' output:\n${err.message}`);
                 return resolve();
             }
