+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pablo",
-  "version": "2.0.3",
+  "version": "2.0.4",
   "description": "Just a useful bot.",
   "main": "index.js",
   "scripts": {
@@ -25,7 +25,7 @@
     "croner": "^10.0.1",
     "discord.js": "^14.25.1",
     "pino": "^10.3.1",
-    "undici": "^7.24.1"
+    "undici": "^7.24.2"
   },
   "optionalDependencies": {
     "bufferutil": "^4.1.0",
@@ -33,7 +33,7 @@
     "zlib-sync": "^0.1.10"
   },
   "devDependencies": {
-    "@biomejs/biome": "^2.4.6",
+    "@biomejs/biome": "^2.4.7",
     "blocked-at": "^1.2.0"
   },
   "repository": {
